ureq_proto::parser

Function try_parse_request

pub fn try_parse_request<const N: usize>(
    input: &[u8],
) -> Result<Option<(usize, Request<()>)>, Error>

Parse bytes into a complete request.

Complete means that the last HTTP header is followed by an \r\n.

If the result is None, the bytes did not contain a full request. That typically means you need to read more bytes and append to the in input buffer before trying again.

The first usize in the resulting pair, is the number of bytes required from the input buffer to form the request.

The const N is the number of headers to max expect. If the input has more headers than N you get an error Error::HttpParseTooManyHeaders.